abstruct and create episode component and adjust some styles
