+++ components/AudioPlayer.module.css
@@ -2,28 +2,22 @@
 	position: relative;
 	width: 100%;
 	height: 100%;
-	/* @@ updated */
-	padding: 0 3rem;
 	display: grid;
 	grid-template-columns: 100%;
-	grid-template-rows: 10% 30% 45% 15%;
+	grid-template-rows: 30% 50% 20%;
 	justify-content: center;
 	align-items: center;
-	/* @@ updated */
 }
 
 .image_wrapper {
-	/* @@ updated */
 	position: relative;
 	width: 60%;
 	height: 90%;
-	/* @@ updated */
 	margin: 1rem auto;
 	box-shadow: 0 0 20px black;
 }
 
 .episode_info_section {
-	/* @@ updated */
 	height: 100%;
 	/* align-self: flex-start; */
 	align-self: center;
@@ -32,33 +26,26 @@
 	grid-template-rows: auto auto;
 	gap: 1rem;
 	padding-bottom: 1.5rem;
-	/* @@ updated */
 }
 
 .episode_title {
 	text-align: center;
-	/* @@ updated */
 	margin: 1rem 0;
 	display: flex;
 	flex-direction: column;
 	justify-content: center;
 	align-items: center;
 	gap: 0.5rem;
-	/* @@ updated */
 }
 
-/* @@ updated */
 .episode_title p {
 	font-weight: 600;
 	margin: 0;
 }
-/* @@ updated */
 
 .episode_details {
-	/* @@ updated */
 	position: relative;
-	padding: 0 1rem;
-	/* @@ updated */
+	padding: 1rem 2rem;
 	overflow: scroll;
 	text-align: start;
 }
@@ -67,25 +54,26 @@
 	display: none;
 }
 
-/* @@ updated */
 .episode_details p {
+	line-height: 2.2rem;
 	margin: 0;
 }
-/* @@ updated */
 
 .audioPlayer {
 	--primary: grey;
 	--secondary: white;
-	/* @@ updated */
 	text-align: center;
 	align-self: flex-start;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
 }
 
 .audioPlayerPlaybutton {
 	display: flex;
 	justify-content: center;
-	gap: 20px;
-	/* @@ updated */
+	align-items: center;
+	gap: 2rem;
 	margin-bottom: 1rem;
 }
 
@@ -124,8 +112,7 @@
 .audioPlayerBar {
 	display: flex;
 	justify-content: center;
-	align-items: center;
-	gap: 20px;
+	gap: 2rem;
 }
 
 .currentTime,
@@ -147,6 +134,7 @@
 	width: 50%;
 	height: 6px;
 	outline: none;
+	margin-top: 0.5rem;
 }
 
 /* progress bar - safari */
