+++ components/episodes.module.css
@@ -1,0 +1,11 @@
+.episodes_section_container {
+	position: relative;
+	width: 100%;
+	height: 100%;
+	padding: 0 3rem;
+	display: grid;
+	grid-template-columns: 100%;
+	grid-template-rows: 10% 90%;
+	justify-content: center;
+	align-items: center;
+}
