+++ pages/index.js
@@ -1,11 +1,6 @@
 import Head from 'next/head';
+import Episodes from '../components/episodes';
 import About from '../components/about';
-import AudioPlayer from '../components/AudioPlayer';
-import AliceCarousel from 'react-alice-carousel';
-import 'react-alice-carousel/lib/alice-carousel.css';
-
-// eslint-disable-next-line react/jsx-key
-const items = [<AudioPlayer />, <AudioPlayer />, <AudioPlayer />];
 
 export default function Home() {
 	return (
@@ -15,7 +10,7 @@ export default function Home() {
 				<meta name="description" content="Generated by create next app" />
 				<link rel="icon" href="/favicon.ico" />
 			</Head>
-			<AliceCarousel mouseTracking items={items} controlsStrategy="alternate" />
+			<Episodes />
 			<About />
 		</div>
 	);
